requirements/development.txt
# Packages required by the application in the development environment.

# Packages specified in base.txt should be installed first.
-r base.txt

django-debug-toolbar==2.2
astroid==2.3.3
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.4.4
six==1.14.0
typed-ast==1.4.1
wrapt==1.11.2
pylint-django==2.0.14
pylint-plugin-utils==0.6
